+++ src/components/AdminDisplays/AdminPlayerDisplay.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { DeletePlayer } from '../../api/DeletePlayer';
+import { DeletePlayer } from '../../api/Player';
 import { TrashIcon, PencilAltIcon } from '@heroicons/react/solid';
 
 const AdminPlayerDisplay = ({player}) => {
 
     function clickHandler(){
-        DeletePlayer(player.playerId);
+        if(window.confirm("This change is permanate, are you sure you want to delete?")) DeletePlayer(player.playerId);
     }
     
     return(
@@ -19,7 +19,7 @@ const AdminPlayerDisplay = ({player}) => {
                 <td>{player.positionId}</td>
                 {player.isActive?<td>Active</td>:<td>Retired</td>}
                 <td>
-                    <Link to = {"/EditPlayer?id=" + player.playerId}><PencilAltIcon className={"inline h-10 w-10 text-purple-500"}/></Link>
+                    <Link to = {"/Admin/EditPlayer?id=" + player.playerId}><PencilAltIcon className={"inline h-10 w-10 text-purple-500"}/></Link>
                     <button onClick = {clickHandler}><TrashIcon className={"inline h-10 w-10 text-purple-500"}/></button>
                 </td>
             </tr>
